Move preprocessors into utils
