+++ autoemulate/experimental/emulators/base.py
@@ -5,8 +5,11 @@ import numpy as np
 from sklearn.base import BaseEstimator
 from torch import nn, optim
 
-from autoemulate.experimental.data.preprocessors import Preprocessor
-from autoemulate.experimental.data.utils import ConversionMixin, ValidationMixin
+from autoemulate.experimental.data.utils import (
+    ConversionMixin,
+    Preprocessor,
+    ValidationMixin,
+)
 from autoemulate.experimental.types import NumpyLike, OutputLike, TensorLike, TuneConfig
