+++ autoemulate/experimental/emulators/gaussian_process/exact.py
@@ -23,7 +23,7 @@ from autoemulate.emulators.gaussian_process import (
     rq_kernel,
     zero_mean,
 )
-from autoemulate.experimental.data.preprocessors import Preprocessor, Standardizer
+from autoemulate.experimental.data.utils import Preprocessor, Standardizer
 from autoemulate.experimental.emulators.base import Emulator
 from autoemulate.experimental.emulators.gaussian_process import (
     CovarModuleFn,
